requirements.txt
-e git+https://github.com/django-nonrel/django-nonrel.git#egg=django
-e git+https://github.com/emperorcezar/djangotoolbox.git@develop#egg=djangotoolbox
-e git+https://github.com/emperorcezar/mongodb-engine.git#egg=mongodb-engine
-e git+https://github.com/django-nonrel/django-permission-backend-nonrel.git#egg=permission_backend_nonrel
-e git+https://github.com/andresdouglas/django-tastypie.git@nonrel#egg=tastypie
django-tinymce==1.5.1a1
mimeparse==0.1.3
python-dateutil==1.5
python-digest==1.7
# wsgiref==0.1.2
pil
django_compressor
